a first, working model
Based on previous work and some initial design ideas, this adds a very
simplistic framework for showing photos. It creates one, statically
configured frame with one panel in it for showing photos. The
displayed photo is changed every 2 seconds (on delay, not interval) by
a controller that will scale to control multiple frames and panels in
the future. The sequence of photos comes from simply scanning a given
directory for photos. There's no inherent order or
randomization. There are controls for adding more frames and more
panels to a frame, but they have no listeners attached yet.