persist frame state across app restarts
On the first startup, you want a default frame to show up. After
you've adjusted the frame to your liking, you probably want it to show
up where you last had it.

This saves the state of the frame(s) on shutdown so that when you
start the app again, they come back where you had them.

Obviously it's untested with multiple frames because there's only the
one frame right now, but theoretically you could modify the state file
manually...

Note after the fact: I copy/pasted a second frame config in the state
file, and it worked as expected.