fix a few controller issues
Just a few things I noticed while running the app and seeing
issues.

The frame is unadopted from the controller by the App. The
controller doesn't have to do it.

I'm not sure how the "frames.panels"
thing even worked. Maybe groovy magic?

The photo change scheduler was set for a 2-second initial delay. I
don't know why. It just made it take longer for the first photo to
show up.

The SynchronousQueue was from before I realized I just needed to
submit a Callable and get the future.

The rest is informational.