make the executors an App-level resource
The App is meant to be a central class for anything to reach global
resources. The ExecutorServices created here are a good example of
that. The controller should access them through the App, not have them
as member fields itself. (Note this flies in the face of dependency
injection, but I'm not doing a lot of testing here...)