handle panels with wrongly sized images
An action that resizes a panel messes with the visual display of the
photos. This is a first step to make panels show photos as accurately
as possible. If a panel has a photo that's sized correctly, it's drawn
as is. Otherwise, it scales the image during the paint() call. The
current scaling is dumb and distorts the image. Next step is to scale
it correctly, but cheaply.