add rating-based randomization
This is a really big change that adds a second loading stage to enable
fuller features. The first, simple phase just gets the basic
visualization up with a basic PhotoRotation. Then more processing
happens to detect all the photos, load metadata about them, and come
up with a better PhotoRotation that chooses photos weighted by
rating. When finished, it swaps the new one into the photo controller
in place of the original, simple one.

It's still really raw, with a lot of the new code directly in the App
class, and the new PhotoRotation doesn't know how to handle running
out of photos of a given rating. Overall, it proves the general
approach works, which is using a sqlite db to manage picking random
photos.