add first hotkey for quitting the app
Hotkeys and popups menus are to come, but for now, it's nice to have a
key to quit the app, especially when I've set it to "disctraction
free" in the config file, meaning there are no window controls on it.

Also moves the autochange stuff from a fixed schedule to a timer task
so that I can cancel and restart it when I want.