beef up sqlite randomizer
The sqlite randomization algorithm is working really well. This fixes
some small glitches with it, like the broken "end of cycle" image and
the fact that I was cycling wrong by updating the db with the new
cycle instead of just moving to the new cycle in memory. It seems
pretty solid now and includes a nifty logger to show the state of each
ratings cycle on each call.