auto-rotate photos to correct orientation
Images have metadata embedded showing the correct orientation. We can
inspect the metadata and show the images in the correct orientation
here, saving me from having to fix the orientation for all of them
manually.

Mixed in here is also a fix that handles some race or failure
conditions where a panel state shows it needs work but the work keeps
getting cancelled.  One way this could happen is if the same image
gets assigned to a panel twice in a row. The state would show "new
photo assigned", but when the loader goes to work on it, it would see
that the panel already has the photo and cancel the work. The state
wasn't being updated, though, so the work would get restarted over and
over.