focus the file system listing
The "random walk" photo rotation replaced the simple "file system"
rotation. That one's no longer needed. In addition, I noticed that the
file system lister is restricing itself to only files with a .jpg
extension (ignoring case). This is the problem I tried to solve in the
"random walk" rotation by actually probing each file for its content
type, but I don't think that's a good idea here, where I use this
class to build the initial database of photos. It would probably slow
it down tremendously. I should test that and see. For now, this
expands the list of file name extensions it allows.