add support for rewrite files
Some photos benefit from touchups, like adjusting color levels, but I
don't really want to edit the originals, so this adds the ability to
"rewrite" a photo by creating a file with the same name with
"-rewrite" appended (before the extension). The idea is that you can
take an original photo, edit it, and save it as <original
path>-rewrite.<original extension> and it'll be loaded in place of the
original. You'll never otherwise see the original. From the
perspective of the app, the two files are the same photo.

That means the rewrites have to be excluded when listing photos, and
when loading a photo, you have to look for a rewrite file for it. I
think this does that.