don't try to select for a rating that doesn't exist
The sqlite-based random selection was picking a rating from "rating
frequencies" collection without checking whether any photos actually
existed for that rating. That ended in a broken "Cycle Exhausted"
image showing up any time the non-existent rating was picked. This
fixes the class to only pick from ratings that exist in the db.