fix photo root settings
Instead of having to enter the photo path every time, now the settings
for the photo root dir work. Setting an env var to a path will load
photos from that path. In addition, the photo-db defaults to being
relative to the photo dir, so you can have multiple, separate
libraries.