now able to change photo ratings
The popup menu was already added, but this feature wasn't
implemented. Now the controls that change photo ratings work, and you
can save the resulting changes to the db with a new hotkey.

There was a problem with the popup menu. It's being built too early,
when the first panels are initially constructed, and the photo db
(data source) hasn't been loaded at that time due to the "quick
startup". The panels are already up and showing photos before the full
database has been loaded. That's by design so that the UI can appear
quickly, before all data has been loaded and the proper photo sequence
has been calculated, which takes some time.