add photo frequency metrics
This adds charts showing the frequency at which photos are shown by
both parent dir and rating. In adding this, I realized that the
default for graphite reporting is to convert metric names to camel
case, and I had to figure out how to disable that so that I could
wildcard match on the dir name and rating parts of the names.

It occurs to me that a dir name isn't a great candidate to be part of
a metric name. This is something that a tag would be much better
for. I better start investigating tagged metrics.