fix rating change
The first pass at changing a photo's rating only changed it in the
local photo data. That means it didn't affect the running app. You had
to save and restart for the change to have an effect. This makes
rating changes update both the data and the current running state.