handle clicks with no photo
When first starting the app, I'd sometimes see errors because I
accidentally clicked in the frame before a photo had been delivered to
the panel there. It's because all the code assumes a panel has a
photo, which is almost always true, except at startup. This just adds
a null check to return on the popup listener, which is the most likely
place for this to happen.