add animated gif handling
Animated gif's are tricky and have to be done totally differently than
single images. At least, the way I found to do it is different. It
makes sense, since everything so far has consisted of loading one,
still image, and a gif contains multiple images. This adds a working
implementation with some code duplication. It at least works, even if
it's not the best possible implementation.