fix decoration toggle
Before this, toggling window decoration on and off had a strange
effect: the panels in the frame shrunk each time you decorated or
undecorated it. I traced it to this code, which assumes it's always
setting up a new frame but is reused when you make major changes to
the frame, like setting decoration on or off. To do things like that,
the old JFrame is destroyed, and a new one is built, but the state of
the PhotoFrame itself remains.

In this case, the list of PhotoPanels associated with the PhotoFrame
still had the panels left over from the previous frame. Every
decorate/undecorate caused more panels to be added to the list, but
only the most recent ones were being updated and painted.

This may not be the right way to solve this. I need to think about the
fact that a JFrame's lifecycle is separate from a PhotoFrame's.