add full screen
Lots of fullscreen support already exists in the code, centered on the
FrameState, which tracks normal and fullscreen configs. This adds a
hotkey to toggle between fullscreen and not. It seems to work fairly
well for a first try. I just had to fix one place in the PhotoFrame,
where it saves the current frame size based on resize events. It was
always saving to the "normal" config instead of whatever the current
config is.