hackily handle the jframe dispose problem
A dispose listener is added to the PhotoFrame so that I can shut down
the app when the last frame is closed. These listeners get connected
to a corresponding JFrame listener. The problem is that the JFrame can
be disposed and recreated when the PhotoFrame needs to be fullscreened
or undecorated, for example. That causes bad events to fire and wrong
things to happen, like making the photo controller unadopt all the
frames and panels!

This fixes it by disabling the dispose listeners briefly when an
underlying JFrame needs to be recreated. I'm not convinced this is the
right approach. I need to think more on the nature of the two things:
the PhotoFrame vs the JFrame.