fix gif rewrite checks
I saw a rewrite inconsistency in the logs like:

```
15:37:46.853 W - Found a rewrite inconsistency. Cheap path determined 2020-05-14/DSCF3131.gif but comprehensive path determined 2020-05-14\DSCF3131-rewrite.JPG - App [general-worker-3].
```

I don't know what the base file was because it wasn't getting logged
(fixed), but I wrote some tests for the rewrite behavior and found the
problem: the comprehensive method was ignoring the file extension and
just taking whatever it found. Fixed it to check for the correct
extension.