fix bug with rewrites from previous commit
I fixed one issue with rewrites in the previous commit but introduced a
bigger one: the "comprehensive" check wasn't comparing things correctly
because of the lowercase call. All of the test cases used files with all
lowercase names, so they missed it.

This adds more test cases with some uppercase names, for which I had to
refactor the test. Instead of a static set of files for every test, now
every test can specify the files it wants. It's much more flexible now
and will be able to mimic any new situation I come across that needs to
be tested.