don't duplicate action listeners
I'm recognizing that the PhotoFrame has things mixed up. There are two
pieces there: the PhotoFrame, representing one logical window that
lasts a long time, and the JFrame which is the visible window that
could be disposed and re-created multiple times over the life of a
PhotoFrame. One thing that's wrong is that the action listeners are
added to the buttons every time a JFrame is created, meaning that
toggling distraction-free on and off will add more listeners, and then
buttons start firing multiple times on a single click. This fixes that.