directly manage panel add/remove
The controller was losing track of panels when changing
configurations, especially when switching between fullscreen and
not. It was because a change like that entails disposing the existing
JFrame and building a new one, including all new panels, and the
controller wasn't kept up to date with the changes. This removes some
of the more complicated code that was trying to keep things in sync
and simply ensures the controller is notified each time a panel is
created or destroyed. Since it's only done in one place, this should
be a simple fix.