don't manage frames in controller
The controller doesn't really need to know about frames. Originally, I
think I intended for it to know about the panels it needed to manage
by looking at the frames. The previous commit made that idea obsolete
because the frames now notify the controller directly of every new and
removed panel. This simplifies things some by only keeping a list of
frames in the App, which seems like where it belongs. It was there
already, so having another list in the controller was just
duplication.